requirements.txt
pyspark
delta-spark
pandas
sqlalchemy
psycopg2-binary